% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_colwiseinverse.R
\name{argmin_between_two}
\alias{argmin_between_two}
\title{Find the argmin between two propositions}
\usage{
argmin_between_two(m, ind, A, prop1, prop2)
}
\arguments{
\item{m}{The column whose a cell need to be updated.}

\item{ind}{The number of the cell in the column.}

\item{A}{A square matrix to column-wise inverse.}

\item{prop1}{First proposition for the \code{ind} cell in \code{m}.}

\item{prop2}{Second proposition for the \code{ind} cell in \code{m}}
}
\value{
\code{prop1} or \code{prop2}.
}
\description{
Find the argmin between two propositions
}
\details{
The function to minimize is \eqn{m'Am}.
}
