% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftconf.R
\name{as_shiftconf}
\alias{as_shiftconf}
\alias{print.shiftconf}
\alias{plot.shiftconf}
\title{'shiftconf' object}
\usage{
as_shiftconf(x, shiftestim)

\method{print}{shiftconf}(x, digits = 3, ...)

\method{plot}{shiftconf}(x, digits = 3, ...)
}
\arguments{
\item{x}{a 'shiftconf' object.}

\item{shiftestim}{a \code{shiftestim} object.}

\item{digits}{Number of digits to round to.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
a \code{shiftconf} object.
}
\description{
'shiftconf' object
}
