% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftestim.R
\name{as_shiftestim}
\alias{as_shiftestim}
\alias{print.shiftestim}
\alias{plot.shiftestim}
\title{'shiftestim' object}
\usage{
as_shiftestim(listopt, tree, zscores, lambda, alphaOU)

\method{print}{shiftestim}(x, digits = 3, ...)

\method{plot}{shiftestim}(x, digits = 3, ...)
}
\arguments{
\item{listopt}{an output of 'estimate_shifts()'}

\item{tree}{tree}

\item{zscores}{vector of observed z-scores}

\item{lambda}{regularization parameter}

\item{alphaOU}{parameter}

\item{x}{a 'shiftestim' object.}

\item{digits}{Number of digits to round to.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
a 'shiftestim' object
}
\description{
'shiftestim' object
}
