% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_desparsified.R
\name{beta}
\alias{beta}
\title{Compute beta}
\usage{
beta(X, y, beta_init, score_system)
}
\arguments{
\item{X}{A vector of size m*(n+m).}

\item{y}{A vector of size m.}

\item{beta_init}{Initial value of beta found with scaled lasso.}

\item{score_system}{Score system of X}
}
\value{
The one-step self-bias corrected estimator of beta.
}
\description{
Compute beta
}
