% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\alias{loglikelihood}
\title{Bayesian information criterion}
\usage{
bic(obs_zscores, est_zscores, est_shifts, sigma)

loglikelihood(obs_zscores, est_zscores, sigma)
}
\arguments{
\item{obs_zscores}{observed z-scores}

\item{est_zscores}{estimated z-scores}

\item{est_shifts}{estimated z-scores}

\item{sigma}{standard error}
}
\value{
scalar
}
\description{
Bayesian information criterion
}
\examples{
bic(obs_zscores = c(-2.7, -2.3, 0.2, -0.1, -2, 0.1),
    est_zscores = c(-2.5, -2.5, 0, 0, -1, 0),
    est_shifts = c(0, 0, 0, 0, -1, 0, -2.5, 0, 0, 0),
    sigma = 1.4)
}
