% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_system.R
\name{calculate_Z}
\alias{calculate_Z}
\title{Score system}
\usage{
calculate_Z(X)
}
\arguments{
\item{X}{Matrix to pseudo orthogonalize of size m*(n+m).}
}
\value{
The matrix of score system, same dimension as X.
}
\description{
Compute the score system of a matrix.
}
