% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_scoresystem.R
\name{calculate_Z_column_with_best_lambda}
\alias{calculate_Z_column_with_best_lambda}
\title{Compute a column of the score system}
\usage{
calculate_Z_column_with_best_lambda(i, X, lambda)
}
\arguments{
\item{i}{Integer. The column to use.}

\item{X}{Matrix to pseudo orthogonalize of size m*(n+m).}

\item{lambda}{Numeric.}
}
\value{
A column of the score system.
}
\description{
Compute a column of the score system
}
