% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_scoresystem.R
\name{calculate_Z_with_best_lambda}
\alias{calculate_Z_with_best_lambda}
\title{Compute the score system once lambda is chosen}
\usage{
calculate_Z_with_best_lambda(X, lambda)
}
\arguments{
\item{X}{Matrix to pseudo orthogonalize of size m*(n+m).}

\item{lambda}{Numeric.}
}
\value{
The score system.
}
\description{
Compute the score system once lambda is chosen
}
