% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{check_selection}
\alias{check_selection}
\title{Check if a shiftestim came from model selection}
\usage{
check_selection(shiftestim)
}
\arguments{
\item{shiftestim}{a shiftestim object.}
}
\value{
logical.
}
\description{
Check if a shiftestim came from model selection
}
