% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_system.R
\name{choose_best_lambda}
\alias{choose_best_lambda}
\title{Choose best lambda by cross-validation}
\usage{
choose_best_lambda(lambdas, X)
}
\arguments{
\item{lambdas}{Numeric vector of lambda to test.}

\item{X}{Matrix to pseudo orthogonalize of size m*(n+m).}
}
\value{
The best lambda, which minimize RMSE prediction
}
\description{
Choose best lambda by cross-validation
}
