% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_colwiseinverse.R
\name{coef_p2}
\alias{coef_p2}
\title{Compute polynomial coefficients.}
\usage{
coef_p2(m, ind, A)
}
\arguments{
\item{m}{The column whose a cell need to be updated.}

\item{ind}{The number of the cell in the column.}

\item{A}{A semi positive definite matrix}
}
\value{
A list with \code{a}, \code{b} and \code{c} components.
}
\description{
Compute coefficients of the polynomial \eqn{m'Am} with the cell
\code{ind} of \code{m} as the indeterminate.
}
\examples{
  m <- c(-1, 4)
  A <- matrix(c(3, 2, 2, 5), nrow = 2)
  coef_p2(A = A, m = m, ind = 1)
  coef_p2(A = A, m = m, ind = 2)
}
