% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{compute_gradient_function}
\alias{compute_gradient_function}
\title{Compute gradient function}
\usage{
compute_gradient_function(Y, X, lambda)
}
\arguments{
\item{Y}{A vector of size m.}

\item{X}{A vector of size m*(n+m).}

\item{lambda}{a grid of positive regularization parameters}
}
\value{
a function that take Delta (a vector of size n+m) as argument
and returns a vector of size n+m
}
\description{
Compute gradient function
}
