% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_colwiseinverse.R
\name{confint_colwiseinverse}
\alias{confint_colwiseinverse}
\title{Desparsified lasso based on column-wise inverse matrix}
\usage{
confint_colwiseinverse(x, alpha_conf = 0.05, ...)
}
\arguments{
\item{x}{a 'shiftestim' object.}

\item{alpha_conf}{the confidence level.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
a list
}
\description{
Desparsified lasso based on column-wise inverse matrix
}
