% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_desparsified.R
\name{confint_desparsified}
\alias{confint_desparsified}
\title{Compute score system based confidence intervals}
\usage{
confint_desparsified(x, alpha_conf = 0.05, ...)
}
\arguments{
\item{x}{a 'shiftestim' object.}

\item{alpha_conf}{the confidence level.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
a list
}
\description{
Compute score system based confidence intervals
}
