% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_scoresystem.R
\name{confint_scoresystem}
\alias{confint_scoresystem}
\title{Desparsified lasso based on score system}
\usage{
confint_scoresystem(x, alpha_conf = 0.05, ...)
}
\arguments{
\item{x}{a 'shiftpunct' object.}

\item{alpha_conf}{the confidence level.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
a list
}
\description{
Desparsified lasso based on score system
}
