% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{correct_pvalues}
\alias{correct_pvalues}
\title{Correct p-values}
\usage{
correct_pvalues(pv, alpha)
}
\arguments{
\item{pv}{Vector of p-values to correct}

\item{alpha}{Threshold of detection}
}
\value{
The corrected p-values
}
\description{
Correct p-values
}
\details{
Correction is applicable at a fixed threshold.
}
