% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_desparsified.R
\name{covariance_noise_matrix}
\alias{covariance_noise_matrix}
\title{Covariance noise matrix}
\usage{
covariance_noise_matrix(X, score_system)
}
\arguments{
\item{X}{Matrix size m*(n+m).}

\item{score_system}{Score system of X.}
}
\value{
The covariance of the noise component for the leaves (size n*n)
}
\description{
Covariance noise matrix
}
