% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_colwiseinverse.R,
%   R/confint_scoresystem.R
\name{covariance_noise_matrix_colwiseinverse}
\alias{covariance_noise_matrix_colwiseinverse}
\alias{covariance_noise_matrix_scoresystem}
\title{Covariance noise matrix}
\usage{
covariance_noise_matrix_colwiseinverse(X, colwiseinverse, XTXn)

covariance_noise_matrix_scoresystem(X, score_system)
}
\arguments{
\item{X}{Matrix size m*(n+m).}

\item{colwiseinverse}{The columnwise inverse.}

\item{XTXn}{Matrix.}

\item{score_system}{Score system of \code{X}.}
}
\value{
The covariance of the noise component for the leaves (size n*n)
}
\description{
Covariance noise matrix
}
