% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{create_clusters}
\alias{create_clusters}
\title{Create clusters}
\usage{
create_clusters(
  tree,
  N_clusters,
  method = c("monophyletic", "paraphyletic", "uniform")
)
}
\arguments{
\item{tree}{tree whose tips must be clustered.}

\item{N_clusters}{number of clusters.}

\item{method}{method to create clusters among \code{"monophyletic"},
\code{"paraphyletic"} and \code{"uniform"}.}
}
\value{
The belonging of each tip to a cluster as a named integer.
}
\description{
Create clusters
}
\examples{
tree <- ape::rtree(10)
N <- 3
plot(tree)
create_clusters(tree, N)
create_clusters(tree, N, method = "paraphyletic")
}
