% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{create_listopt_pms}
\alias{create_listopt_pms}
\title{Create a list from}
\usage{
create_listopt_pms(objective_value, shift_est, method, best_alpha,
  best_lambda)
}
\arguments{
\item{objective_value}{the objective value.}

\item{shift_est}{the estimated shifts.}

\item{method}{the used method.}

\item{best_alpha}{the best alpha during the model selection phase.}

\item{best_lambda}{the best lambda during the model selection phase.}
}
\value{
a list.
}
\description{
Create a list from
}
