% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftconf_utils.R
\name{df_confint_pvalue}
\alias{df_confint_pvalue}
\title{Confidence interval and p-values for estimates}
\usage{
df_confint_pvalue(estimate, sigma, tau, alpha_conf = 0.05)
}
\arguments{
\item{estimate}{Punctual estimate. Eventually named}

\item{sigma}{Associated standard error, length \code{1}.}

\item{tau}{Noise factor, same length as \code{estimate}.}

\item{alpha_conf}{Confidence level.}
}
\value{
A dataframe with 4 (or 5 if \code{estimate} has names) columns.
}
\description{
From a vector of estimates computes the lower and upper bound and a p-value.
}
\details{
The confidence interval is bivariate whereas the p-value is univariate.
}
