% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_global.R
\name{estimate_shifts}
\alias{estimate_shifts}
\title{Estimate shifts}
\usage{
estimate_shifts(
  Delta0,
  zscores,
  tree,
  alphaOU,
  lambda = NULL,
  method = c("L-BFGS-B", "shooting", "scaledlasso"),
  criterion = c("bic", "pbic"),
  constraint_type = c("beta", "yhat", "none"),
  ...
)
}
\arguments{
\item{Delta0}{initial position (size n+m)}

\item{zscores}{z-scores (size m)}

\item{tree}{the tree used to compute the incidence and covariance matrices
(if not specified)}

\item{alphaOU}{a grid of positive alpha parameters}

\item{lambda}{a grid of positive regularization parameters}

\item{method}{method to use for the optimization. One of \code{L-BFGS-B}
or \code{shooting}.}

\item{criterion}{criterion on which the selection is done.}

\item{constraint_type}{Constrains on shifts.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an object of class \code{shiftestim}
}
\description{
Estimate shifts
}
