% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{estimate_tstar}
\alias{estimate_tstar}
\title{Estimate threshold of rejection for t-scores}
\usage{
estimate_tstar(t_scores, alpha)
}
\arguments{
\item{t_scores}{Vector of t-scores.}

\item{alpha}{Threshold of detection.}
}
\value{
The threshold of rejection for t-scores.
}
\description{
Estimate threshold of rejection for t-scores
}
