% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_leaves.R
\name{extract_from_cidf}
\alias{extract_from_cidf}
\title{Extract significant indices}
\usage{
extract_from_cidf(cidf, side = c("left", "both", "right"))
}
\arguments{
\item{cidf}{confidence interval dataframe with \code{"estimate"},
\code{"lower"}, \code{"upper"} columns.}

\item{side}{The side where z-scores are significant (\code{"left"}
(default), \code{"both"} or \code{"right"}).}
}
\description{
Extract indices where lower and upper components of the dataframe
are both positive or both negative.
}
