% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_models}
\alias{extract_models}
\title{Extract models computed during selection}
\usage{
extract_models(shiftpunct)
}
\arguments{
\item{shiftpunct}{a shiftpunct object.}
}
\value{
a list of shiftpunct objects.
}
\description{
Extract models computed during selection
}
