% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftconf_utils.R
\name{extract_significant_leaves}
\alias{extract_significant_leaves}
\title{Extract significant leafs}
\usage{
extract_significant_leaves(x, side = c("left", "both", "right"))
}
\arguments{
\item{x}{a 'shiftestim' object.}

\item{side}{The side where z-scores are significant (\code{"left"}
(default), \code{"both"} or \code{"right"}). Used when method is
\code{"desparsified lasso"}.}
}
\value{
The names of the significant leafs
}
\description{
Extract significant leafs
}
