% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_leaves.R
\name{extract_significant_leaves.shiftconf}
\alias{extract_significant_leaves.shiftconf}
\title{Extract significant leafs}
\usage{
\method{extract_significant_leaves}{shiftconf}(x, threshold = 0.05, ...)
}
\arguments{
\item{x}{A "shiftconf" object.}

\item{threshold}{P-value threshold.}

\item{...}{
  Arguments passed on to \code{\link[=extract_significant_leaves]{extract_significant_leaves}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
The names of the significant leafs.
}
\description{
Extract significant leafs
}
