% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_leaves.R
\name{extract_significant_leaves.shiftpunct}
\alias{extract_significant_leaves.shiftpunct}
\title{Extract significant leafs}
\usage{
\method{extract_significant_leaves}{shiftpunct}(
  x,
  threshold = 0,
  direction = c("<", "<=", ">", ">="),
  ...
)
}
\arguments{
\item{x}{A "shiftpunct" object.}

\item{threshold}{The threshold where selection begins.}

\item{direction}{The direction of the selection from the threshold.
Default to \code{"<"}.}

\item{...}{
  Arguments passed on to \code{\link[=extract_significant_leaves]{extract_significant_leaves}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
The names of the significant leafs
}
\description{
Extract significant leafs
}
