% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_helpers.R
\name{fitch}
\alias{fitch}
\title{Fitch parsimony score}
\usage{
fitch(phy, states)
}
\arguments{
\item{phy}{tree}

\item{states}{named state vector (typically the estimated zscores)}
}
\value{
Fitch parsimony score
}
\description{
Fitch parsimony score
}
\examples{
tree <- read.tree(text = "(((t4:0.07418493367,(t6:0.0466889315,t5:0.0466889315):0.02749600216):0.6619201357,t2:0.7361050694):0.4046613234,(t7:0.3876200329,(t1:0.09560661687,t3:0.09560661687):0.2920134161):0.7531463598);")
states <- c(t1 = "A", t2 = "B", t3 = "A", t4 = "C", t5 = "C", t6 = "C", t7 = "A")
fitch(tree, states) ## 2
}
