% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_helpers.R
\name{force_ultrametric}
\alias{force_ultrametric}
\title{Force a tree to be ultrametric}
\usage{
force_ultrametric(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
an ultrametric tree
}
\description{
Force a tree to be ultrametric by extending its terminal branches
}
\examples{
tree_u <- force_ultrametric(alcohol$tree)
ape::is.ultrametric(tree_u)
}
