% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hyperparameters.R
\name{generate_gamma}
\alias{generate_gamma}
\title{Gamma}
\usage{
generate_gamma(X, factor = 2)
}
\arguments{
\item{X}{Design matrix, size \code{m*(n+m)}.}

\item{factor}{Numeric.}
}
\value{
A suggested value of \eqn{\gamma}.
}
\description{
Generate a gamma for the column-wise inverse procedure.
}
