% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_system.R
\name{get_lambda_sequence}
\alias{get_lambda_sequence}
\title{Vector of lambdas to test}
\usage{
get_lambda_sequence(X)
}
\arguments{
\item{X}{Matrix to pseudo orthogonalize of size m*(n+m).}
}
\value{
100 values for lambdas
}
\description{
Vector of lambdas to test
}
