% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix.R
\name{incidence_matrix}
\alias{incidence_matrix}
\title{Incidence matrix}
\usage{
incidence_matrix(phylo)
}
\arguments{
\item{phylo}{a tree}
}
\value{
the incidence matrix
}
\description{
Compute the incidence matrix from a tree
}
\examples{
tree <- ape::rtree(3)
incidence_matrix(tree)
}
