% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{inverse_sqrt}
\alias{inverse_sqrt}
\title{Inverse of the square root of a matrix}
\usage{
inverse_sqrt(M)
}
\arguments{
\item{M}{a symmetric positive definite matrix}
}
\value{
a lower triangular matrix
}
\description{
Compute \eqn{R}, the inverse of the square root of a matrix \eqn{M} such that \deqn{M^{-1} = R'R.}
}
\examples{
M <- matrix(2, ncol = 3, nrow = 3) + diag(5, ncol = 3, nrow = 3)
R <- inverse_sqrt(M)
t(R) \%*\% R \%*\% M ## Should be the Identity
}
