% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{lambda_grid}
\alias{lambda_grid}
\title{Helper function to compute a logarithmic lambda grid}
\usage{
lambda_grid(x, y, n_lambda = 6, min_ratio = 1e-05)
}
\arguments{
\item{x}{Design matrix}

\item{y}{Response vector}

\item{n_lambda}{Length of the lambda grid}

\item{min_ratio}{Ratio between the minimum and maximum values in the lambda grid}
}
\value{
a numeric vector of length \code{nlambda}
}
\description{
Helper function to compute a logarithmic lambda grid
}
