% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_colwiseinverse.R
\name{minimun_p2}
\alias{minimun_p2}
\title{Compute the minimum of a second order polynomial}
\usage{
minimun_p2(coef)
}
\arguments{
\item{coef}{A list with \code{a}, \code{b} and \code{c} components.}
}
\value{
A numeric of length one (eventually \code{Inf} or \code{-Inf})
if the polynomial if convex. \code{c(-Inf, Inf)} if the polynomial
is strictly concave.
}
\description{
Compute the minimum of a second order polynomial
}
\examples{
minimun_p2(list(a = 1, b = 2, c = 0))
minimun_p2(list(a = -1, b = 2, c = 0))
minimun_p2(list(a = 0, b = 2, c = 0))
minimun_p2(list(a = 0, b = 0, c = 3))
}
\seealso{
\code{\link{coef_p2}}
}
