% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_desparsified.R
\name{noise_factor}
\alias{noise_factor}
\title{Noise factor}
\usage{
noise_factor(X, score_system)
}
\arguments{
\item{X}{Matrix size m*(n+m).}

\item{score_system}{Score system of X.}
}
\value{
The vector of noise factor for the shifts, size (n+m)
}
\description{
Noise factor
}
