% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_utils.R
\name{noise_factor_scoresystem}
\alias{noise_factor_scoresystem}
\alias{noise_factor_colwiseinverse}
\title{Noise factor}
\usage{
noise_factor_scoresystem(X, score_system)

noise_factor_colwiseinverse(X, colwiseinverse, XTXn)
}
\arguments{
\item{X}{Matrix size m*(n+m).}

\item{score_system}{Score system of \code{X}.}

\item{colwiseinverse}{The columnwise inverse.}

\item{XTXn}{Matrix.}
}
\value{
The vector of noise factor for the shifts, size (n+m)
}
\description{
Noise factor
}
