% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscores.R
\name{p2z}
\alias{p2z}
\title{Transform p-values to z-scores}
\usage{
p2z(p.value, e.sign = NULL, tol = 1e-15)
}
\arguments{
\item{p.value}{a vector of p-values}

\item{e.sign}{a vector of signs of the effects, taken on value -1 and 1}

\item{tol}{a numeric value at which the p-value (both ends) will truncate}
}
\value{
a vector of z-values
}
\description{
Transform p-values to z-scores
}
\examples{
pv <- runif(5)
p2z(pv)
p2z(0.5) ## should be 0
}
