% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_helpers.R
\name{parsimony_score}
\alias{parsimony_score}
\title{Fitch parsimony score}
\usage{
parsimony_score(tree, zscores)
}
\arguments{
\item{tree}{tree}

\item{zscores}{named z-scores}
}
\value{
Fitch parsimony score
}
\description{
Fitch parsimony score
}
