% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shifts.R
\name{plot_shifts}
\alias{plot_shifts}
\title{Plot shift amplitude and location on the tree}
\usage{
plot_shifts(tree, shifts, true_scores = NULL, obs_scores = NULL,
  est_scores = NULL, sup_scores = NULL, digits = 3)
}
\arguments{
\item{tree}{Phylo-class object}

\item{shifts}{Branch-associated shifts vector, sorted in cladewise order}

\item{true_scores}{Leaf-associated scores, as a named vector.}

\item{obs_scores}{Leaf-associated scores, as a named vector.}

\item{est_scores}{Leaf-associated scores, as a named vector.}

\item{sup_scores}{A list of lists for additional scores to plot. Each
list must contain at least a 'scores" element and eventually a
'title" and 'color' one.}

\item{digits}{Number of digits to round to.}
}
\value{
a \code{ggplot} object, as created by \code{ggtree}
}
\description{
Plot
}
\details{
If scores is unnamed, assumes scores are in the same order as
\code{tree$tip.label}
}
\examples{
tree <- ape::read.tree(text = "(((A,B),C),(D,E));")
shifts <- c(0, -3, 0, 0, 0, 0, -2, 0)
scores <- c(-3.2, -2.8, 0.1, -2.1, -0.1)
plot_shifts(tree, shifts, obs_scores = scores)
}
