% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pvalues.R
\name{pull_pvalues.shiftconf}
\alias{pull_pvalues.shiftconf}
\title{Pull p-values}
\usage{
\method{pull_pvalues}{shiftconf}(x)
}
\arguments{
\item{x}{Object of class shiftestim or shiftconf}
}
\value{
The named vector of p-values
}
\description{
Pull p-values
}
