% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_scoresystem.R
\name{rescale_score_system}
\alias{rescale_score_system}
\title{Rescale the score system}
\usage{
rescale_score_system(Z, X)
}
\arguments{
\item{Z}{The non-scaled score system.}

\item{X}{Matrix to pseudo orthogonalize of size m*(n+m).}
}
\value{
The scaled score system.
}
\description{
Rescale the score system
}
