% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_scoresystem.R
\name{rmse_glmnet}
\alias{rmse_glmnet}
\title{Compute prediction errors}
\usage{
rmse_glmnet(c, K, dataselects, X, lambdas)
}
\arguments{
\item{c}{Column to use.}

\item{K}{Number of batches.}

\item{dataselects}{Batches.}

\item{X}{Matrix to pseudo orthogonalize of size m*(n+m).}

\item{lambdas}{Numeric vector of lambda to test.}
}
\value{
Matrix of errors per lambda and batch.
}
\description{
Compute prediction errors
}
