% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_scaled.R
\name{scaled_lasso}
\alias{scaled_lasso}
\title{Scaled lasso}
\usage{
scaled_lasso(y, X, projected = TRUE, ...)
}
\arguments{
\item{y}{A vector of size m.}

\item{X}{A vector of size m*(n+m).}

\item{projected}{Logical. Should each coordinate of \code{beta_init} be projected on \eqn{\mathbb{R}_{-}}?}

\item{...}{Not used, here for avoiding errors when passing unknown arguments
from previous ellipsis.}
}
\value{
A list composed by \code{beta_init} and \code{hsigma}.
}
\description{
Scaled lasso to compute \code{beta_init} and \code{hsigma}.
}
