% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_desparsified.R
\name{score_system}
\alias{score_system}
\title{Score system}
\usage{
score_system(X, y, beta_init, hsigma)
}
\arguments{
\item{X}{A vector of size m*(n+m).}

\item{y}{A vector of size m.}

\item{beta_init}{Initial value of beta found with scaled lasso.}

\item{hsigma}{Estimate value of sigma, found with scaled lasso.}
}
\value{
The matrix of score system, same dimension as X.
}
\description{
Compute the score system of a
}
