% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simu_zscores}
\alias{simu_zscores}
\title{Simulation of z-scores on branches}
\usage{
simu_zscores(tree, alpha, shifts = NULL, Nshifts = NULL)
}
\arguments{
\item{tree}{tree.}

\item{alpha}{the parameter for the OU process.}

\item{shifts}{vector of shifts.}

\item{Nshifts}{number of shifts. Must be specified if \code{shifts} is not.}
}
\value{
a vector of simulated z-scores.
}
\description{
Simulates z-scores
}
\details{
Shifts are sampled from "statistically significant values of z-scores"
}
\examples{
tree <- ape::rcoal(7)
shift = c(0, 0, -3, 0, 0, 0, 0, 0, 0, -2, 0, 0)
simu_zscores(tree, alpha = 0.2, shift = c(0, 0, -3, 0, 0, 0, 0, 0, 0, -2, 0, 0))

}
