% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_desparsified.R
\name{size_half_confint}
\alias{size_half_confint}
\title{Size of the half confidence interval}
\usage{
size_half_confint(noise_factor, hsigma, alpha = 0.05)
}
\arguments{
\item{noise_factor}{Noise factor, size m.}

\item{hsigma}{Estimate value of sigma, found with scaled lasso.}

\item{alpha}{The confidence level.}
}
\value{
The half-size of the confidence interval for each beta, size m.
}
\description{
Compute the size of the half confidence interval, depending on the desired
level of confidence \code{alpha}.
}
