% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_scoresystem.R
\name{size_half_confint_shifts}
\alias{size_half_confint_shifts}
\title{Size of the half confidence interval for the shifts}
\usage{
size_half_confint_shifts(noise_factor, hsigma, alpha_conf = 0.05)
}
\arguments{
\item{noise_factor}{Noise factor, size m.}

\item{hsigma}{Estimate value of sigma, found with scaled lasso.}

\item{alpha_conf}{The confidence level.}
}
\value{
The half-size of the confidence interval for each beta, size m+n.
}
\description{
Compute the size of the half confidence interval for the shifts,
depending on the desired level of confidence \code{alpha}.
}
