% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_desparsified.R
\name{size_half_confint_zscores}
\alias{size_half_confint_zscores}
\title{Size of the half confidence interval for the z-scores}
\usage{
size_half_confint_zscores(
  covariance_noise_mat,
  incidence_mat,
  hsigma,
  alpha_conf = 0.05
)
}
\arguments{
\item{covariance_noise_mat}{The covariance of the noise component
for the leaves.}

\item{incidence_mat}{The incidence matrix of the tree, size \code{n*(n+m)}.}

\item{hsigma}{Estimate value of sigma, found with scaled lasso.}

\item{alpha_conf}{The confidence level.}
}
\value{
The half-size of the confidence interval for each z-score, size n.
}
\description{
Compute the size of the half confidence interval for the z-scores,
depending on the desired level of confidence \code{alpha}.
}
