% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_pvalues.R
\name{smooth_pvalues}
\alias{smooth_pvalues}
\title{Hierarchical correction of p-values}
\usage{
smooth_pvalues(
  pvalues,
  esign = NULL,
  tree,
  arg_shiftpunct = list(alphaOU = 1, method = "scaled lasso"),
  arg_shiftconf = list(alpha_conf = 0.05, method = "score system")
)
}
\arguments{
\item{pvalues}{Vector of p-values.}

\item{esign}{Vector of e-signs (optional).}

\item{tree}{Tree, as a \code{phylo} object.}

\item{arg_shiftpunct}{Arguments to be passed to
\code{estimate_shifts} function.}

\item{arg_shiftconf}{Arguments to be passed to
\code{estimate_confint} function.}
}
\value{
A corrected vector.
}
\description{
Hierarchical correction of p-values
}
\examples{
pval_obs <- test_kruskalwallis(chlamydiae$X, chlamydiae$Y)$p.value
tree <- force_ultrametric(chlamydiae$tree)
smooth_pvalues(pvalue = pval_obs, tree = tree,
               arg_shiftpunct = list(alpha = c(0.1, 2),
                                     method = "scaled lasso"))
}
