% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_colwiseinverse.R
\name{solve_colwiseinverse_col}
\alias{solve_colwiseinverse_col}
\title{Compute a column for the column-wise inverse.}
\usage{
solve_colwiseinverse_col(col, A, gamma)
}
\arguments{
\item{col}{The column number.}

\item{A}{A square matrix to column-wise inverse.}

\item{gamma}{Non-negative.}
}
\value{
The column, wile respecting constrains.
}
\description{
Compute a column for the column-wise inverse.
}
