% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_colwiseinverse.R
\name{solve_colwiseinverse_col_cell}
\alias{solve_colwiseinverse_col_cell}
\title{Update column cell}
\usage{
solve_colwiseinverse_col_cell(m, ind, col, A, gamma)
}
\arguments{
\item{m}{The column whose a cell need to be updated.}

\item{ind}{The number of the cell in the column.}

\item{col}{The column number.}

\item{A}{A square matrix to column-wise inverse.}

\item{gamma}{Non-negative.}
}
\value{
The updated column.
}
\description{
Update column cell
}
