% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_desparsified.R
\name{solve_desparsified}
\alias{solve_desparsified}
\title{Desparsified lasso}
\usage{
solve_desparsified(Delta0, Y, X, alpha_conf = 0.05, ...)
}
\arguments{
\item{Delta0}{Initial value for Delta.}

\item{Y}{A vector of size m.}

\item{X}{A vector of size m*(n+m).}

\item{alpha_conf}{Confidence level.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list composed by \code{par} (beta estimate) among others.
}
\description{
Desparsified lasso
}
