% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_tests.R
\name{test_wilcoxon}
\alias{test_wilcoxon}
\alias{test_kruskalwallis}
\alias{test_fisher}
\title{Vectorized tests}
\usage{
test_wilcoxon(X, Y)

test_kruskalwallis(X, Y)

test_fisher(X, Y)
}
\arguments{
\item{X}{matrix with samples in columns and taxa in rows}

\item{Y}{vector of condition}
}
\value{
A list with a \code{p.value} component and eventually
a \code{e.sign} component.
}
\description{
Vectorized tests
}
\examples{
head(test_wilcoxon(alcohol$X, alcohol$Y)$p.value)
head(test_kruskalwallis(alcohol$X, alcohol$Y)$p.value)
head(test_fisher(alcohol$X, alcohol$Y)$p.value)
}
