% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_helpers.R
\name{tree_height}
\alias{tree_height}
\title{Tree height}
\usage{
tree_height(tree)
}
\arguments{
\item{tree}{an ultrametric tree \code{phylo} object.}
}
\value{
the height of the tree
}
\description{
Compute the tree height from an ultrametric tree
}
\examples{
tree <- ape::rcoal(6)
tree_height(tree)
}
