% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_scoresystem.R
\name{update_beta}
\alias{update_beta}
\title{Update beta}
\usage{
update_beta(X, y, beta_init, score_system)
}
\arguments{
\item{X}{Matrix size m*(n+m).}

\item{y}{A vector of size m.}

\item{beta_init}{Initial value of beta found with scaled lasso.}

\item{score_system}{Score system of X.}
}
\value{
The one-step self-bias corrected estimator of beta, size m.
}
\description{
Update beta
}
