% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_utils.R
\name{update_beta_scoresystem}
\alias{update_beta_scoresystem}
\alias{update_beta_colwiseinverse}
\title{Update beta}
\usage{
update_beta_scoresystem(X, y, beta_init, score_system)

update_beta_colwiseinverse(X, y, beta_init, colwiseinverse)
}
\arguments{
\item{X}{Matrix size m*(n+m).}

\item{y}{A vector of size m.}

\item{beta_init}{Initial value of beta found with scaled lasso.}

\item{score_system}{Score system of \code{X}.}

\item{colwiseinverse}{The columnwise inverse.}
}
\value{
The one-step self-bias corrected estimator of beta, size m.
}
\description{
Update beta
}
