% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_utils.R
\name{update_confint}
\alias{update_confint}
\title{Update confidence interval}
\usage{
update_confint(x, alpha_confint)
}
\arguments{
\item{x}{a 'shiftestim' object.}

\item{alpha_confint}{Confidence level.}
}
\value{
a 'shiftestim' object
}
\description{
Update confidence interval
}
