% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_scaled.R
\name{update_sigma}
\alias{update_sigma}
\title{Update sigma in the scaled lasso algorithm}
\usage{
update_sigma(y, X, beta)
}
\arguments{
\item{y}{A vector of size m.}

\item{X}{A vector of size m*(n+m).}

\item{beta}{The current value of beta.}
}
\value{
Updated version of sigma.
}
\description{
Compute the exact value of sigma that nullify the gradient.
}
