% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zazou-package.R
\docType{package}
\name{zazou-package}
\alias{zazou}
\alias{zazou-package}
\title{zazou: Z-Scores As Ornstein-Uhlenbeck}
\description{
Differentially abundance analysis with Ornstein-Uhlenbeck
    process on z-scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abichat/zazou}
  \item Report bugs at \url{https://github.com/abichat/zazou/issues}
}

}
\author{
\strong{Maintainer}: Antoine Bichat \email{a.bichat@yahoo.fr} (\href{https://orcid.org/0000-0001-6599-7081}{ORCID})

Authors:
\itemize{
  \item Mahendra Mariadassou \email{mahendra.mariadassou@inra.fr} (\href{https://orcid.org/0000-0003-2986-354X}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
