% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\alias{chlamydiae}
\title{Alcohol data set}
\format{A list containing
\describe{
  \item{\code{X}}{the normalized OTU data,}
  \item{\code{Y}}{the alcohol intake phenotype,}
  \item{\code{tree}}{the corresponding phylogenetic tree,}
  \item{\code{otu.name}}{the taxonomic lineages of the OTUs.}
}}
\source{
Wu, Gary D., et al. "Linking long-term dietary patterns with gut
microbial enterotypes." Science 334.6052 (2011): 105-108.
}
\usage{
alcohol

chlamydiae
}
\description{
The data set was taken from a study of the dietary association with the gut
microbiome (Wu, 2011). Here we include the alcohol intake as the primary
phenotype. The OTU abundance data consists of 98 samples with 949 OTUs
with prevalence > 10%. The raw counts are normalized to account for
variable library sizes.
}
\keyword{datasets}
