% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{bic}
\alias{bic}
\alias{loglikelihood}
\alias{pbic}
\title{Information criteria}
\usage{
bic(obs_zscores, est_zscores, est_shifts, sigma)

loglikelihood(obs_zscores, est_zscores, sigma)

pbic(obs_zscores, est_zscores, est_shifts, sigma, alpha, tree)
}
\arguments{
\item{obs_zscores}{observed z-scores}

\item{est_zscores}{estimated z-scores}

\item{est_shifts}{estimated z-scores}

\item{sigma}{standard error}

\item{alpha}{alpha from OU.}

\item{tree}{phylogenetic tree.}
}
\value{
scalar
}
\description{
Loglikelihood, Bayesian Information Criterion and Phylogenetic BIC
}
\examples{
bic(obs_zscores = c(-2.7, -2.3, 0.2, -0.1, -2, 0.1),
    est_zscores = c(-2.5, -2.5, 0, 0, -1, 0),
    est_shifts = c(0, 0, 0, 0, -1, 0, -2.5, 0, 0, 0),
    sigma = 1.4)
}
