% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{bic}
\alias{bic}
\alias{loglikelihood}
\alias{pbic}
\title{Information criteria}
\usage{
bic(obs_zscores, est_zscores, est_shifts, mat_covar)

loglikelihood(obs_zscores, est_zscores, mat_covar)

pbic(obs_zscores, est_zscores, est_shifts, mat_incidence, mat_covar)
}
\arguments{
\item{obs_zscores}{Observed z-scores.}

\item{est_zscores}{Estimated z-scores.}

\item{est_shifts}{Estimated z-scores.}

\item{mat_covar}{Covariance matrix.}

\item{mat_incidence}{Incidence matrix.}
}
\value{
scalar
}
\description{
Loglikelihood, Bayesian Information Criterion and Phylogenetic BIC
}
\examples{
bic(obs_zscores = c(-2.7, -2.3, 0.2, -0.1, -2, 0.1),
    est_zscores = c(-2.5, -2.5, 0, 0, -1, 0),
    est_shifts = c(0, 0, 0, 0, -1, 0, -2.5, 0, 0, 0),
    mat_covar = diag(rep(1, 6)))
}
