% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{check_selection}
\alias{check_selection}
\title{Check if a shiftpunct came from model selection}
\usage{
check_selection(shiftpunct)
}
\arguments{
\item{shiftpunct}{a shiftpunct object.}
}
\value{
logical.
}
\description{
Check if a shiftpunct came from model selection
}
