% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_colwiseinverse.R
\name{compute_bounds}
\alias{compute_bounds}
\title{Compute constrains bounds}
\usage{
compute_bounds(m, ind, col, A, gamma)
}
\arguments{
\item{m}{The column whose a cell need to be updated.}

\item{ind}{The number of the cell in the column.}

\item{col}{The column number.}

\item{A}{A square matrix to column-wise inverse.}

\item{gamma}{Non-negative.}
}
\value{
A list with \code{lower_bound} and \code{upper_bound} components.
}
\description{
Compute constrains bounds
}
