% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{compute_objective_function}
\alias{compute_objective_function}
\title{Compute objective function}
\usage{
compute_objective_function(Y, X, lambda)
}
\arguments{
\item{Y}{a vector of size m}

\item{X}{a vector of size m*(n+m)}

\item{lambda}{a grid of positive regularization parameters}
}
\value{
a function that take Delta (a vector of size n+m) as argument
and returns a scalar
}
\description{
Compute objective function
}
