% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{compute_objective_function}
\alias{compute_objective_function}
\title{Compute objective function}
\usage{
compute_objective_function(
  Y,
  X,
  lambda,
  sigma = NULL,
  type = c("lasso", "scaledlasso")
)
}
\arguments{
\item{Y}{A vector of size m.}

\item{X}{A vector of size m*(n+m).}

\item{lambda}{a grid of positive regularization parameters}

\item{sigma}{Needed if \code{type} is set to \code{"scaledlasso"}.}

\item{type}{Character. If you want the \code{"lasso"} or
\code{"scaledlasso"} objective function.}
}
\value{
a function that take Delta (a vector of size n+m) as argument
and returns a scalar
}
\description{
Compute objective function
}
