% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_colwiseinverse.R
\name{confint_colwiseinverse}
\alias{confint_colwiseinverse}
\title{Desparsified lasso based on column-wise inverse matrix}
\usage{
confint_colwiseinverse(x, alpha_conf = 0.05, silent_on_errors = TRUE, ...)
}
\arguments{
\item{x}{A 'shiftestim' object.}

\item{alpha_conf}{The confidence level.}

\item{silent_on_errors}{Logical, default to TRUE.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
a list
}
\description{
Desparsified lasso based on column-wise inverse matrix
}
