% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_pvalues.R
\name{correct_pvalues}
\alias{correct_pvalues}
\title{Hierarchical correction of p-values}
\usage{
correct_pvalues(
  pvalues,
  esign = NULL,
  tree,
  arg_shiftestim = list(alphaOU = 1, method = "scaledlasso"),
  arg_shiftconf = list(alpha_conf = 0.05, method = "scoresystem")
)
}
\arguments{
\item{pvalues}{Vector of p-values.}

\item{esign}{Vector of e-signs (optional).}

\item{tree}{Tree, as a \code{phylo} object.}

\item{arg_shiftestim}{Arguments to be passed to
\code{estimate_shifts} function.}

\item{arg_shiftconf}{Arguments to be passed to
\code{estimate_confint} function.}
}
\value{
A corrected vector.
}
\description{
Hierarchical correction of p-values
}
\examples{
pval_obs <- test_kruskalwallis(chlamydiae$X, chlamydiae$Y)$p.value
tree <- force_ultrametric(chlamydiae$tree)
correct_pvalues(pvalue = pval_obs, tree = tree,
                arg_shiftestim = list(alpha = c(0.1, 2),
                                      method = "scaledlasso"))
}
