% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_global.R
\name{estimate_confint}
\alias{estimate_confint}
\title{Estimate confidence interval from scaled lasso.}
\usage{
estimate_confint(
  shiftestim,
  alpha_conf = 0.05,
  method = c("scoresystem", "colwiseinverse"),
  ...
)
}
\arguments{
\item{shiftestim}{a \code{shiftestim} object with method \code{scaledlasso}.}

\item{alpha_conf}{the confidence level.}

\item{method}{For the moment, only \code{scoresystem}.}

\item{...}{further argument...}
}
\value{
a \code{shiftconf} object.
}
\description{
Estimate confidence interval from scaled lasso.
}
