% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_global.R
\name{estimate_shifts}
\alias{estimate_shifts}
\title{Estimate shifts}
\usage{
estimate_shifts(Delta0, zscores, incidence_mat, covar_mat, lambda)
}
\arguments{
\item{Delta0}{initial position (size n+m)}

\item{zscores}{z-scores (size m)}

\item{incidence_mat}{incidence matrix (size m*(n+m))}

\item{covar_mat}{covariance matrix (size m*m)}

\item{lambda}{a positive regularization parameter}
}
\value{
a list
}
\description{
Estimate shifts
}
