% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftconf_utils.R
\name{extract_from_cidf}
\alias{extract_from_cidf}
\title{Extract significant indices}
\usage{
extract_from_cidf(cidf, side = c("left", "both", "right"))
}
\arguments{
\item{cidf}{dataframe with \code{"estimate"}, \code{"lower"},
\code{"upper"} columns.}

\item{side}{The side where z-scores are significant (\code{"left"}
(default), \code{"both"} or \code{"right"}). Used when method is
\code{"desparsified lasso"}.}
}
\description{
Extract indices where lower and upper components of the dataframe
are both positive or both negative.
}
