% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_models}
\alias{extract_models}
\title{Extract models computed during selection}
\usage{
extract_models(shiftestim)
}
\arguments{
\item{shiftestim}{a shiftestim object.}
}
\value{
a list of shiftestim objects.
}
\description{
Extract models computed during selection
}
