% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_leaves.R
\name{extract_significant_leaves}
\alias{extract_significant_leaves}
\alias{extract_significant_leaves.NULL}
\title{Extract significant leafs}
\usage{
extract_significant_leaves(x, ...)

\method{extract_significant_leaves}{`NULL`}(x, ...)
}
\arguments{
\item{x}{Object of class}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
The names of the significant leafs.
}
\description{
Extract significant leafs
}
\seealso{
\code{\link{extract_significant_leaves.shiftestim}},
\code{\link{extract_significant_leaves.shiftconf}}
}
