% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{lambda_grid}
\alias{lambda_grid}
\title{Helper function to compute a logarithmic lambda grid}
\usage{
lambda_grid(x, y, n_lambda = 10, min_ratio = 0.01, ...)
}
\arguments{
\item{x}{Design matrix.}

\item{y}{Response vector.}

\item{n_lambda}{Length of the lambda grid.}

\item{min_ratio}{Ratio between the minimum and maximum values in
the lambda grid.}

\item{...}{Not used, here for avoiding errors when passing unknow arguments
from previous ellipsis.}
}
\value{
a numeric vector of length \code{nlambda}
}
\description{
Helper function to compute a logarithmic lambda grid
}
