% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{norm0}
\alias{norm0}
\title{Norm 0}
\usage{
norm0(x)
}
\arguments{
\item{x}{Numeric.}
}
\value{
The number of non-zero values in x (\eqn{\|x\|_0}).
}
\description{
Compute the number of non-zero values
}
\examples{
x <- c(1, 2, 3, 0, 4)
norm0(x)
}
