% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{norm0}
\alias{norm0}
\title{Norm 0}
\usage{
norm0(x, reverse = FALSE)
}
\arguments{
\item{x}{Numeric.}

\item{reverse}{If \code{TRUE}, returns the number of zeros in \code{x}.}
}
\value{
By default, the number of non-zero values in \code{x}
(\eqn{\|x\|_0}). If \code{reverse} is set to \code{TRUE},
the number of zero in \code{x}.
}
\description{
Compute the number of non-zero values
}
\examples{
x <- c(1, 2, 3, 0, 4)
norm0(x)
}
