% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shifts.R
\name{plot_shifts}
\alias{plot_shifts}
\title{Plot shift amplitude and location on the tree}
\usage{
plot_shifts(shifts, tree, scores = NULL)
}
\arguments{
\item{shifts}{Branch-associated shifts vector, sorted in cladewise order}

\item{tree}{Phylo-class object}

\item{scores}{Leaf-associated scores. Either a named vector}
}
\value{
a \code{ggplot} object, as created by \code{ggtree}
}
\description{
Plot shift amplitude and location on the tree
}
\examples{
tree <- ape::read.tree(text = "(((A,B),C),(D,E));")
shifts <- c(0, -2.9513814680442, 0, 0, 0, -0.00195465052163029, -1.896800199013, 0)
scores <- c(-3, -3, 0, -2, 0)
plot_shifts(shifts, tree, scores)
}
