% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_scaled.R
\name{scaled_lasso}
\alias{scaled_lasso}
\title{Scaled lasso}
\usage{
scaled_lasso(beta0, y, X, lambda = NULL, ...)
}
\arguments{
\item{beta0}{Initial value for beta.}

\item{y}{A vector of size m.}

\item{X}{A vector of size m*(n+m).}

\item{lambda}{A regularization parameter. If \code{NULL}, a universal
value is chosen.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list composed by \code{par} (beta estimate) and
\code{sigma_scaledlasso} among others.
}
\description{
Scaled lasso to compute \code{beta_init} and \code{hsigma}.
}
