% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simu_zscores}
\alias{simu_zscores}
\title{Simulation of z-scores on branches}
\usage{
simu_zscores(tree, alpha, shifts = NULL, Nshifts = NULL, graph = FALSE, ...)
}
\arguments{
\item{tree}{tree.}

\item{alpha}{the parameter for the OU process.}

\item{shifts}{vector of shifts.}

\item{Nshifts}{number of shifts. Must be specified if \code{shifts} is not.}

\item{graph}{show the tree, shifts and z-scores be plotted?}

\item{...}{additional parameters to pass to \code{plot_shifts()}.}
}
\value{
a vector of simulated z-scores.
}
\description{
Simulates z-scores
}
\details{
Shifts are sampled from "statistically significant values of z-scores"
}
\examples{
set.seed(42)
tree <- ape::rcoal(5)
shift <- c(0, 0, -3, 0, 0, 0, 0, 0)
simu_zscores(tree, alpha = 0.2, shift = shift, graph = TRUE)
simu_zscores(tree, alpha = 0.2, Nshifts = 2, graph = TRUE)
}
