% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_colwiseinverse.R
\name{solve_colwiseinverse}
\alias{solve_colwiseinverse}
\title{Solve column-wise inverse}
\usage{
solve_colwiseinverse(A, gamma)
}
\arguments{
\item{A}{A square matrix to column-wise inverse.}

\item{gamma}{Non-negative.}
}
\value{
The column-wise, inverse, same size as \code{A}.
}
\description{
Solve column-wise inverse
}
