% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_colwiseinverse.R
\name{solve_colwiseinverse}
\alias{solve_colwiseinverse}
\title{Solve column-wise inverse}
\usage{
solve_colwiseinverse(
  A,
  gamma,
  ntry_max = 5e+05,
  silent_on_tries = TRUE,
  silent_on_errors = TRUE,
  ...
)
}
\arguments{
\item{A}{A semi positive definite matrix}

\item{gamma}{Numeric. Non-negative}

\item{ntry_max}{Integer. Maximum umber of try for each column.}

\item{silent_on_errors}{Logical, default to TRUE.}

\item{...}{Further arguments to be passed to or from other methods.}

\item{silent_on_try}{Logical, default to TRUE.}
}
\value{
The column-wise, inverse, same size as \code{A}.
}
\description{
Solve column-wise inverse
}
