% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_colwiseinverse.R
\name{solve_colwiseinverse_col}
\alias{solve_colwiseinverse_col}
\title{Compute a column for the column-wise inverse.}
\usage{
solve_colwiseinverse_col(col, A, gamma, m0, max_it = 5000, ...)
}
\arguments{
\item{col}{The column number.}

\item{A}{A semi positive definite matrix}

\item{gamma}{Numeric. Non-negative}

\item{m0}{Startup column.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
The column, wile respecting constrains.
}
\description{
Compute a column for the column-wise inverse.
}
