% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_internal.R
\name{solve_univariate}
\alias{solve_univariate}
\title{Solve unidirectional constrained problem}
\usage{
solve_univariate(y, x, lambda = 0)
}
\arguments{
\item{y}{a vector of size n}

\item{x}{a vector of size n}

\item{lambda}{a positive regularization parameter}
}
\value{
The scalar solution of the 1D optimization problem
}
\description{
This function minimizes \eqn{\beta} in the 1D problem
\eqn{1/2 * ||y - x \beta||_2^2 + \lambda |\beta|} subject to \eqn{\beta < 0}.
}
\details{
The analytical solution of this problem is given by
\deqn{\beta* = min(0, (y'x + \lambda) / x'x ).}
}
\examples{
solve_univariate(1:4, -(4:1), 2)
}
