% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_shooting.R
\name{solve_univariate}
\alias{solve_univariate}
\alias{solve_multivariate}
\title{Solve unidirectional constrained problem}
\usage{
solve_univariate(
  y,
  x,
  z = rep(0, length(y)),
  lambda = 0,
  constraint_type = c("beta", "yhat", "none"),
  ...
)

solve_multivariate(beta0, y, X, lambda, prob = NULL, max_it = 500, ...)
}
\arguments{
\item{y}{a vector of size n.}

\item{x}{a vector of size n.}

\item{z}{a vector of size n.}

\item{lambda}{a grid of positive regularization parameters}

\item{constraint_type}{Character. "beta" (default), "yhat" or "none". Ensures that
all coordinates of \eqn{\beta} (for \code{constraint_type = "beta"}) or
\eqn{z + x\beta} (for \code{constraint_type = "yhat"}) are negative.}

\item{...}{further arguments passed to or from other methods.}

\item{beta0}{the initial position of beta.}

\item{X}{a matrix of size  x p.}

\item{prob}{a vector of probability weights for obtaining the coordinates
to be sampled.}

\item{max_it}{Maximum number of iterations.}
}
\value{
The scalar solution \eqn{\beta} of the 1D optimization problem

the estimated value of beta
}
\description{
This function minimizes \eqn{\beta} in the 1D problem
\eqn{1/2 * ||y - z - x \beta||_2^2 + \lambda |\beta|} subject to
either \eqn{\beta <= 0} or \eqn{z + x\beta <= 0} (coordinate wise).
}
\details{
The analytical solution of this problem is given by
\deqn{\beta* = min(0, ((y-z)'x + \lambda) / x'x ).}
when using the first constraint and is slightly more
complex when using the second constraint
(refer to the corresponding vignette)
}
\examples{
solve_univariate(1:4, -(4:1), 2)
}
